phase6: consolidate websocket entrypoints and split health route
