+++ run_server.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
-# run_server.sh - Easy startup script for the eye-tracking server
+# run_server.sh - Easy startup script for the consolidated backend server
 
-echo "🚀 Starting Eye Tracking Server..."
+echo "🚀 Starting Backend Server..."
 echo "=================================="
 
 # Check if virtual environment exists
@@ -12,17 +12,6 @@ if [ ! -d ".venv" ]; then
     echo "   source .venv/bin/activate  # Linux/Mac"
     echo "   pip install -r requirements.txt"
     exit 1
-fi
-
-# Check if model file exists
-if [ ! -f "models/shape_predictor_68_face_landmarks.dat" ]; then
-    echo "⚠️  Model file not found!"
-    echo "   Download from: http://dlib.net/files/shape_predictor_68_face_landmarks.dat.bz2"
-    echo "   Extract to: models/shape_predictor_68_face_landmarks.dat"
-    echo ""
-    echo "   Or use: curl -o models/shape_predictor_68_face_landmarks.dat.bz2 http://dlib.net/files/shape_predictor_68_face_landmarks.dat.bz2"
-    echo "   Then extract the .bz2 file"
-    echo ""
 fi
 
 # Activate virtual environment and start server
@@ -36,7 +25,7 @@ else
 fi
 
 echo "🔧 Starting FastAPI server on http://localhost:8000"
-echo "📊 WebSocket endpoint: ws://localhost:8000/ws"
+echo "📊 WebSocket endpoints: ws://localhost:8000/ws and ws://localhost:8000/ws/gaze/{session_id}"
 echo "📚 API docs: http://localhost:8000/docs"
 echo ""
 echo "Press Ctrl+C to stop the server"
